Add more info to profile and handle logged in profile, more minor changes
