+++ components/Navbar/Navbar.tsx
@@ -9,11 +9,11 @@ import { FaChartLine, FaTrophy } from 'react-icons/fa6';
 import { useRouter } from 'next/router';
 import Link from 'next/link';
 import { FaUser } from 'react-icons/fa';
-import { UserModal } from './UserModal';
+import { convertToSS58 } from '@/utils/convertToSS58';
 
 export const Navbar = () => {
   const [isClient, setIsClient] = useState(false);
-  const { isConnected, chain } = useAccount();
+  const { isConnected, chain, address } = useAccount();
   const { switchChain } = useSwitchChain();
 
   useEffect(() => {
@@ -27,8 +27,6 @@ export const Navbar = () => {
   }, []);
 
   const router = useRouter();
-
-  const [userModalOpened, setUserModalOpened] = useState<boolean>(false);
 
   return (
     <nav className='bg-[#111217] w-full h-[64px]'>
@@ -69,31 +67,20 @@ export const Navbar = () => {
           <div className='hidden md:block'>
             <Image src={banner} alt='banner' width={400} height={60} />
           </div>
-          {isConnected && (
-            <button
+          {isConnected && address && (
+            <Link
               className='p-2 rounded bg-[#191B24]'
-              onClick={() => setUserModalOpened(true)}
+              href={`/profile/${convertToSS58(address)}`}
             >
               <FaUser
-                className={`text-sm hover:text-[#4ECB7D] cursor-pointer  transition`}
+                className={`text-sm hover:text-[#4ECB7D] cursor-pointer  transition ${
+                  router.asPath ===
+                    `/profile/5DTwZmpc9BJZ2c7DbqBGBTxpYwW3oustdZsDfE6A3REAhU1G` &&
+                  'text-[#4ECB7D]'
+                }`}
               />
-            </button>
+            </Link>
           )}
-
-          {/* <button
-            className='p-2 rounded bg-[#191B24]'
-            onClick={() => changeRoute('/')}
-          >
-            <FaUser
-              className={`text-sm hover:text-[#4ECB7D] cursor-pointer  transition ${
-                router.pathname === '/' && 'text-[#4ECB7D]'
-              }`}
-            />
-          </button> */}
-          {/* <MintButton /> */}
-          {/* {isConnected && <CurrentNetworkTab />} */}
-          {/* <UIConfiguration /> */}
-          {/* {isClient && <WalletConnect />} */}
           {isClient && (
             <ConnectButton
               accountStatus={{
@@ -109,10 +96,6 @@ export const Navbar = () => {
           )}
         </div>
       </div>
-      <UserModal
-        isModalOpened={userModalOpened}
-        handleCloseModal={() => setUserModalOpened(false)}
-      />
     </nav>
   );
 };
